+++ src/client/src/pages/settings/admin/userAdministration.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, FC, MouseEvent, useCallback, useEffect, useState } from "react";
+import { ChangeEvent, FC, MouseEvent, useCallback, useContext, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
 import { Checkbox, Chip, Stack, Tooltip } from "@mui/material";
@@ -14,37 +14,31 @@ import {
 import { User, WorkgroupRole } from "../../../api/apiInterfaces.ts";
 import { fetchUsers, updateUser } from "../../../api/user.ts";
 import { useApiRequest } from "../../../hooks/useApiRequest.ts";
+import { useDeleteUserPrompts } from "../../../hooks/useDeleteEntityPrompts.tsx";
 import { muiLocales } from "../../../mui.locales.ts";
 import { TablePaginationActions } from "../../overview/boreholeTable/TablePaginationActions.tsx";
 import { quickFilterStyles } from "./quickfilterStyles.ts";
-import { useDeleteUserPrompts } from "./useDeleteUserPrompts.tsx";
+import { UserAdministrationContext } from "./userAdministrationContext.tsx";
 import { useSharedTableColumns } from "./useSharedTableColumns.tsx";
 
-interface UserTableProps {
-  setSelectedUser: (user: User | null) => void;
-  users: User[];
-  setUsers: (users: User[]) => void;
-}
-
-export const UserTable: FC<UserTableProps> = ({ setSelectedUser, users, setUsers }) => {
+export const UserAdministration: FC = () => {
   const { t, i18n } = useTranslation();
   const [filterModel, setFilterModel] = useState<GridFilterModel>();
-  const [isLoading, setIsLoading] = useState(true);
   const history = useHistory();
   const { callApiWithErrorHandling, callApiWithRollback } = useApiRequest();
   const { statusColumn, getDeleteColumn } = useSharedTableColumns();
-  const { showDeleteWarning } = useDeleteUserPrompts(setSelectedUser, users, setUsers);
+  const { users, setUsers, setSelectedUser, userTableSortModel, setUserTableSortModel } =
+    useContext(UserAdministrationContext);
+  const { showDeleteUserWarning } = useDeleteUserPrompts(setSelectedUser, users, setUsers);
   const handleFilterModelChange = useCallback((newModel: GridFilterModel) => setFilterModel(newModel), []);
 
   useEffect(() => {
-    setIsLoading(true);
+    setSelectedUser(null);
     const getUsers = async () => {
       const users: User[] = await callApiWithErrorHandling(fetchUsers, []);
       setUsers(users);
-      setIsLoading(false);
     };
     getUsers();
-    setSelectedUser(null);
   }, [callApiWithErrorHandling, setSelectedUser, setUsers, t]);
 
   const renderCellCheckbox = (params: GridRenderCellParams) => {
@@ -129,9 +123,7 @@ export const UserTable: FC<UserTableProps> = ({ setSelectedUser, users, setUsers
   const handleDeleteUser = (event: MouseEvent<HTMLButtonElement>, id: number) => {
     event.stopPropagation();
     const user = users.find(user => user.id === id);
-    if (!user) return;
-    setSelectedUser(user);
-    showDeleteWarning(user);
+    showDeleteUserWarning(user);
   };
 
   const columns: GridColDef[] = [
@@ -176,7 +168,7 @@ export const UserTable: FC<UserTableProps> = ({ setSelectedUser, users, setUsers
       getRowClassName={getRowClassName}
       rowHeight={44}
       sortingOrder={["asc", "desc"]}
-      loading={isLoading}
+      loading={!users?.length}
       onRowClick={handleRowClick}
       rowCount={users?.length}
       rows={users}
@@ -203,6 +195,8 @@ export const UserTable: FC<UserTableProps> = ({ setSelectedUser, users, setUsers
       disableDensitySelector
       filterModel={filterModel}
       onFilterModelChange={handleFilterModelChange}
+      sortModel={userTableSortModel}
+      onSortModelChange={setUserTableSortModel}
     />
   );
 };