+++ src/client/src/pages/settings/settingsPage.tsx
@@ -1,17 +1,20 @@
-import { useMemo, useState } from "react";
+import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import { Stack } from "@mui/material";
 import { ReduxRootState, User } from "../../api-lib/ReduxStateInterfaces.ts";
-import { User as UserV2 } from "../../api/apiInterfaces.ts";
 import { theme } from "../../AppTheme.ts";
 import { useAuth } from "../../auth/useBdmsAuth.tsx";
 import { TabPanel } from "../../components/tabs/tabPanel.tsx";
 import AboutSettings from "./aboutSettings";
 import AdminSettings from "./admin/adminSettings";
+import { UserAdministration } from "./admin/userAdministration.tsx";
+import { UserAdministrationProvider } from "./admin/userAdministrationContext.tsx";
 import { UserDetail } from "./admin/userDetail.tsx";
-import { UserTable } from "./admin/userTable.tsx";
+import { WorkgroupAdministration } from "./admin/workgroupAdministration.tsx";
+import { WorkgroupAdministrationProvider } from "./admin/workgroupAdministrationContext.tsx";
+import { WorkgroupDetail } from "./admin/workgroupDetail.tsx";
 import EditorSettings from "./editorSettings.tsx";
 import { SettingsHeader } from "./settingsHeader.tsx";
 import TermSettings from "./termSettings";
@@ -20,9 +23,6 @@ export const SettingsPage = () => {
   const auth = useAuth();
   const { t } = useTranslation();
   const currentUser: User = useSelector((state: ReduxRootState) => state.core_user);
-  const [selectedUser, setSelectedUser] = useState<UserV2 | null>(null);
-  const [users, setUsers] = useState<UserV2[]>([]);
-
   const isAdminUser = currentUser.data.admin;
   const isAnonymousUser = auth.anonymousModeEnabled;
 
@@ -34,44 +34,47 @@ export const SettingsPage = () => {
       tabsArray.push({ label: t("terms"), hash: "terms", component: <TermSettings /> });
     }
     if (isAdminUser) {
-      tabsArray.unshift({ label: t("workgroups"), hash: "workgroups", component: <AdminSettings /> });
+      tabsArray.unshift({ label: t("legacySettings"), hash: "legacysettings", component: <AdminSettings /> });
+      tabsArray.unshift({
+        label: t("workgroups"),
+        hash: "workgroups",
+        component: <WorkgroupAdministration />,
+      });
       tabsArray.unshift({
         label: t("users"),
         hash: "users",
-        component: <UserTable setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />,
+        component: <UserAdministration />,
       });
     }
 
     return tabsArray;
-  }, [isAdminUser, isAnonymousUser, t, users]);
+  }, [isAdminUser, isAnonymousUser, t]);
 
   return (
-    <>
-      <SettingsHeader selectedUser={selectedUser} setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />
-      <Switch>
-        <Route
-          exact={false}
-          key={4}
-          path={"/setting/user/:id"}
-          render={() => <UserDetail user={selectedUser} setUser={setSelectedUser} />}
-        />
-        <Route
-          exact={false}
-          key={4}
-          path={"/setting"}
-          render={() => (
-            <Stack
-              sx={{
-                height: "100%",
-                p: 5,
-                overflowY: "auto",
-                backgroundColor: theme.palette.background.lightgrey,
-              }}>
-              <TabPanel tabs={tabs} />
-            </Stack>
-          )}
-        />
-      </Switch>
-    </>
+    <UserAdministrationProvider>
+      <WorkgroupAdministrationProvider>
+        <SettingsHeader />
+        <Switch>
+          <Route exact={false} key={4} path={"/setting/user/:id"} render={() => <UserDetail />} />
+          <Route exact={false} key={5} path={"/setting/workgroup/:id"} render={() => <WorkgroupDetail />} />
+          <Route
+            exact={false}
+            key={6}
+            path={"/setting"}
+            render={() => (
+              <Stack
+                sx={{
+                  height: "100%",
+                  p: 5,
+                  overflowY: "auto",
+                  backgroundColor: theme.palette.background.lightgrey,
+                }}>
+                <TabPanel tabs={tabs} />
+              </Stack>
+            )}
+          />
+        </Switch>
+      </WorkgroupAdministrationProvider>
+    </UserAdministrationProvider>
   );
 };